refactor server, part 1 (#1116)
This update does not change any behavior.
The simply moves code around into a new directory structure.
Later patches will add new capabilities.
