+++ server/test/test_fbs.py
@@ -4,7 +4,7 @@ import numpy as np
 from scipy import sparse
 
 import decode_fbs
-from server.app.util.fbs.matrix import encode_matrix_fbs, decode_matrix_fbs
+from server.data_common.fbs.matrix import encode_matrix_fbs, decode_matrix_fbs
 
 
 class FbsTests(unittest.TestCase):
