--- server/utils/constants.py
@@ -1,1 +1,0 @@
-MODES = ["umap", "tsne", "draw_graph_fa", "draw_graph_fr", "diffmap", "phate"]
